Frontend Commit 1
This is the first major commit to the front end of this project. With state, users can create new tends through a form, as well as edit the value of tends. With context, the list of tends is provided to every component in the application. Implementation may change as a Firebase back end is developed. Everything implemented so far is responsive.